refactor(filechannel): ChannelService를 Service와 Repository로 분리하고 Factory 수정 - FileChannelService에서 Service와 Repository 계층으로 분리 - FileChannelService는 Repository 주입 방식으로 변경 - FileServiceFactory에서 channelService 생성 부분을 Repository 기반으로 수정
